feat: Add app information bottom sheet with developer contacts
feat: Implement featured section with horizontal scrollable song cards

feat: Create featured song card with enhanced styling and readability

feat: Add floating search button with animation

feat: Implement highlighted text widget for search results

feat: Create home app bar with title and search functionality

feat: Add profile button to show app information

feat: Implement recent lyric item widget for displaying songs

feat: Create simple section header widget for titles

fix: Update lyrics page to use new color scheme methods for opacity

fix: Update onboarding page to use new color scheme methods for opacity

feat: Add main bottom navigation bar for easy navigation between sections
