+++ lib/features/db_sync/page.dart
@@ -129,7 +129,7 @@ class _DatabaseSyncPageState extends State<DatabaseSyncPage> {
     final textTheme = Theme.of(context).textTheme;
 
     return Scaffold(
-      backgroundColor: colorScheme.background,
+      backgroundColor: colorScheme.surface,
       body: Center(
         child: Padding(
           padding: const EdgeInsets.symmetric(horizontal: 32.0),
@@ -145,7 +145,7 @@ class _DatabaseSyncPageState extends State<DatabaseSyncPage> {
                   shape: BoxShape.circle,
                   boxShadow: [
                     BoxShadow(
-                      color: colorScheme.primary.withOpacity(0.2),
+                      color: colorScheme.primary.withValues(alpha: 0.2),
                       blurRadius: 20,
                       offset: const Offset(0, 8),
                     ),
@@ -164,7 +164,7 @@ class _DatabaseSyncPageState extends State<DatabaseSyncPage> {
                 'Setting Up Tononkira',
                 style: textTheme.headlineSmall?.copyWith(
                   fontWeight: FontWeight.bold,
-                  color: colorScheme.onBackground,
+                  color: colorScheme.onSurface,
                 ),
               ),
               const SizedBox(height: 12),
